// src/interfaces/app/transaction.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface Transaction {
    code:    number;
    success: boolean;
    message: string;
    data:    TransactionData[];
}

export interface TransactionData {
    id:              number;
    referenceID:     string;
    proposalNo:      string;
    category:        string;
    customerName:    string;
    fund:            string;
    amount:          number;
    payType:         string;
    payStatus:       string;
    transactionDate: string;
    channel:         string;
}